// GameChanger.Web/ClientApp/src/app/models/teamStat.ts
export interface Root {
  id: string,
  stats_data: TeamStats,
  team_id: string,
}

export interface TeamStats {
  stats: Stats
  players: { [name: string]: Players }
}

export interface Stats {
  defense: Defense
  general: General
  offense: Offense
}
// ...
export interface Players {
  stats: Stats2
}

export interface Stats2 {
  firstName: string
  lastName: string
  number: string
  defense: Defense2
  general: General2
  offense: Offense2
}

export interface Defense2 {
  a: number
  e: number
  h: number
  l: number
  r: number
  w: number
  "#P": number
  "1B": number
  "2B": number
  "3B": number
  "<3": number
  ab: number
  ao: number
  bb: number
  bf: number
  bk: number
  bs: number
  ci: number
  cs: number
  er: number
  fc: number
  gb: number
  go: number
  gs: number
  hr: number
  if: number
  ip: number
  os: number
  po: number
  "S%": number
  sb: number
  sm: number
  so: number
  sv: number
  sw: number
  tb: number
  tc: number
  ts: number
  wp: number
  "<13": number
  "<3%": number
  baa: number
  bbs: number
  "CB%": number
  "CH%": number
  "CT%": number
  "DB%": number
  "DC%": number
  era: number
  "FB%": number
  fip: number
  flb: number
  fly: number
  fps: number
  "GB%": number
  hbp: number
  inp: number
  "K/G": number
  "KB%": number
  "KC%": number
  lnd: number
  lob: number
  loo: number
  "OS%": number
  oss: number
  pik: number
  "RB%": number
  "SB%": number
  "SC%": number
  shb: number
  shf: number
  "SL%": number
  "SM%": number
  sol: number
  svo: number
  "<13%": number
  "CI:C": number
  "CS:C": number
  "DP:P": number
  "FLB%": number
  "FLY%": number
  fpct: number
  "FPS%": number
  fpsh: number
  fpso: number
  fpsw: number
  full: number
  "GP:C": number
  "GP:F": number
  "GP:P": number
  hard: number
  "IC:C": number
  "IP:F": number
  "K/BB": number
  "K/BF": number
  "LND%": number
  "LOB%": number
  lobb: number
  "LOO%": number
  ossm: number
  ossw: number
  "P/BF": number
  "P/IP": number
  "PB:C": number
  "SB:C": number
  "TP:P": number
  weak: number
  whip: number
  outs: number
  babip: number
  "CS:C%": number
  "FPSH%": number
  "FPSO%": number
  "FPSW%": number
  "GO/AO": number
  "HARD%": number
  hrisp: number
  "IP:1B": number
  "IP:2B": number
  "IP:3B": number
  "IP:CF": number
  "IP:LF": number
  "IP:RF": number
  "IP:SF": number
  "IP:SS": number
  lbfpn: number
  lobbs: number
  "PIK:C": number
  "SB:C%": number
  "WEAK%": number
  _0BBINN: number
  _123INN: number
  abrisp: number
  "BB/INN": number
  "outs-C": number
  "outs-P": number
  "outs:C": number
  "outs:F": number
  "123INN%": number
  _1ST2OUT: number
  "BA/RISP": number
  "SBATT:C": number
  "outs-1B": number
  "outs-3B": number
  "1ST2OUT%": number
  _2STRIKES: number
}

export interface General2 {
  gp: number
}

export interface Offense2 {
  h: number
  r: number
  "1B": number
  "2B": number
  "3B": number
  "6+": number
  ab: number
  bb: number
  "C%": number
  ci: number
  cs: number
  fc: number
  gb: number
  gp: number
  hr: number
  ob: number
  os: number
  pa: number
  ps: number
  sb: number
  sm: number
  so: number
  sw: number
  tb: number
  ts: number
  "6+%": number
  avg: number
  "CB%": number
  "CH%": number
  "CT%": number
  "DB%": number
  "DC%": number
  "FB%": number
  flb: number
  "GB%": number
  hbp: number
  inp: number
  "KB%": number
  "KC%": number
  lnd: number
  lob: number
  obp: number
  ops: number
  "OS%": number
  oss: number
  pik: number
  qab: number
  "RB%": number
  rbi: number
  roe: number
  "SB%": number
  "SC%": number
  shb: number
  shf: number
  "SL%": number
  slg: number
  "SM%": number
  sol: number
  "SW%": number
  xbh: number
  "2S+3": number
  "BB/K": number
  "FLB%": number
  full: number
  gidp: number
  gitp: number
  gshr: number
  hard: number
  "LND%": number
  lobb: number
  ossm: number
  ossw: number
  "QAB%": number
  weak: number
  "2S+3%": number
  babip: number
  hrisp: number
  "PA/BB": number
  "PS/PA": number
  abrisp: number
  _2OUTRBI: number
  _3OUTLOB: number
  "BA/RISP": number
  _2STRIKES: number
}